pyproject.toml
[build-system]
requires = ["poetry-core", "cython"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "cmrai"
description = "CMR-AI"
version = "0.1.0"
readme = "README.md"
authors = ["Bayer AG <ivo.baltruschat@bayer.com>"]

[tool.poetry.dependencies]
python = "~3.9"

# --------- pytorch --------- #
torch = "==1.12.1"
torchvision = "==0.13.1"
pytorch-lightning = "==1.8.1"
torchmetrics = "^0.10.0"

# --------- hydra --------- #
hydra-core = "^1.2.0"
hydra-colorlog = "^1.1.0"

# --------- loggers --------- #
wandb = "^0.13.0"

# --------- others --------- #
scikit-learn = "^1.0"
scikit-image = "^0.19.0"
albumentations = "^1.0.3"

orderedattrdict = "^1.6.0"
flatdict = "^4.0.1"
tqdm = "^4.62.2"
natsort = "^7.1.1"
rich = "^12.0.0" # beautiful text formatting in terminal
icecream = "^2.1.1"

mpire = "^2.3.3"
multiprocess = "^0.70.13"

nibabel = "^4.0.1"
pydicom = "^2.3.0"
SimpleITK = "^2.1.1.2"
dcm2niix = "^1.0.20220509"

seaborn = "^0.11.2"
seaborn-image = "^0.5.0"

google-cloud-storage = "^2.4.0"

pyrootutils = "^1.0.4"

monai = "^0.9.1"
numpy = "~1.21.0"
matplotlib = "~3.5.0"
kornia = "^0.6.7"

dump-env = "^1.3" # dump environment to file

h5py = "^3.7.0"
lightning-lite = "1.8.0.post1"
opencv-python-headless = "^4.6.0.66"
zenodo-get = {git = "https://github.com/dvolgyes/zenodo_get"}



[tool.poetry.group.dev.dependencies]
sigpy = "^0.1.23"
jupyterlab = "^3.5.0"
ipympl = "^0.9.2"
itkwidgets = {version = ">=1.0a16", extras = ["lab"]}
ipywidgets = "^8.0.2"
jill = "^0.11.1"
julia = "^0.5.7"

kaggle = "^1.5.12"
tensorflow = "^2.9.1" # tensorflow for model conversion
tf2onnx = "^1.12.0" # tf to onnx conversion
onnx2torch = "^1.5.1" # onnx to torch conversion

# --------- linters --------- #
pre-commit = "^2.15.0" # hooks for applying linters on commit
black = "^22.6.0" # code formatting
isort = "^5.10.1" # import sorting
flake8 = "^4.0.1" # code analysis
nbstripout = "^0.5.0"  # remove output from jupyter notebooks

[tool.pyright]
include = ["cmrai"]
